`refactor` : refactor file name and directory to kebab case amd update inertia render path
