+++ app/Http/Controllers/Auth/PasswordResetLinkController.php
@@ -17,7 +17,7 @@ class PasswordResetLinkController extends Controller
      */
     public function create(): Response
     {
-        return Inertia::render('Auth/ForgotPassword', [
+        return Inertia::render('auth/forgot-password', [
             'status' => session('status'),
         ]);
     }
