+++ app/Http/Controllers/Auth/RegisteredUserController.php
@@ -20,7 +20,9 @@ class RegisteredUserController extends Controller
      */
     public function create(): Response
     {
-        return Inertia::render('Auth/Register');
+        return Inertia::render('auth/register', [
+            'status' => session('status'),
+        ]);
     }
 
     /**
