--- resources/js/Components/InputLabel.tsx
@@ -1,20 +1,0 @@
-import { LabelHTMLAttributes } from 'react';
-
-export default function InputLabel({
-    value,
-    className = '',
-    children,
-    ...props
-}: LabelHTMLAttributes<HTMLLabelElement> & { value?: string }) {
-    return (
-        <label
-            {...props}
-            className={
-                `block text-sm font-medium text-gray-700 ` +
-                className
-            }
-        >
-            {value ? value : children}
-        </label>
-    );
-}
