+++ resources/js/Components/input-error.tsx
@@ -1,0 +1,17 @@
+import type { HTMLAttributes } from "react";
+
+export default function InputError({
+	message,
+	className = "",
+	...props
+}: HTMLAttributes<HTMLParagraphElement> & { message?: string }) {
+	return message ? (
+		<p
+			{...props}
+			// biome-ignore lint/style/useTemplate: <explanation>
+			className={"text-sm text-red-600 " + className}
+		>
+			{message}
+		</p>
+	) : null;
+}
