+++ resources/js/Components/secondary-button.tsx
@@ -1,0 +1,22 @@
+import type { ButtonHTMLAttributes } from "react";
+
+export default function SecondaryButton({
+	type = "button",
+	className = "",
+	disabled,
+	children,
+	...props
+}: ButtonHTMLAttributes<HTMLButtonElement>) {
+	return (
+		<button
+			{...props}
+			type={type}
+			className={`inline-flex items-center rounded-md border border-gray-300 bg-white px-4 py-2 text-xs font-semibold uppercase tracking-widest text-gray-700 shadow-sm transition duration-150 ease-in-out hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2 disabled:opacity-25 ${
+				disabled && "opacity-25"
+			} ${className}`}
+			disabled={disabled}
+		>
+			{children}
+		</button>
+	);
+}
