+++ resources/js/Pages/Auth/forgot-password.tsx
@@ -1,7 +1,7 @@
-import InputError from "@/components/InputError";
-import PrimaryButton from "@/components/PrimaryButton";
-import TextInput from "@/components/TextInput";
-import GuestLayout from "@/layouts/GuestLayout";
+import InputError from "@/components/input-error";
+import PrimaryButton from "@/components/primary-button";
+import TextInput from "@/components/text-input";
+import GuestLayout from "@/layouts/guest-layout";
 import { Head, useForm } from "@inertiajs/react";
 import type { FormEventHandler } from "react";
