+++ resources/js/Pages/Dashboard.tsx
@@ -1,4 +1,4 @@
-import AuthenticatedLayout from "@/layouts/AuthenticatedLayout";
+import AuthenticatedLayout from "@/layouts/dashboard-layout";
 import { Head } from "@inertiajs/react";
 
 export default function Dashboard() {
