--- resources/js/Pages/Welcome.tsx
@@ -1,7 +1,0 @@
-export default function Welcome() {
-	return (
-		<>
-			<p>home</p>
-		</>
-	);
-}
