+++ resources/js/Pages/home.tsx
@@ -1,0 +1,8 @@
+import { useEffect } from "react";
+
+export default function Welcome() {
+	useEffect(() => {
+		window.location.assign("/dashboard");
+	}, []);
+	return null;
+}
